Improve explanation of Turbolinks & jQuery fight
Hey Steven! Your explanation of what was going wrong seemed slightly inaccurate to me. Hopefully this clears it up for you!

Also, `jquery-turbolinks` is fine, but it's probably better-suited to help large projects with lots of existing jQuery logic to adopt Turbolinks. For a small project like this, you could get away without adding an extra dependency.

You have a couple options:

### 1. Hook into one of Turbolinks' [custom events](https://github.com/turbolinks/turbolinks-classic#events), rather than `document ready`:

``` javascript
$(document).on('page:change', function() {
	$(".boat_image").click(function(){
		$(this).siblings(".expedition_form_div").css("display", "inline");
	});
});
```

### 2. Use a global click handler

jQuery lets you "move the handler up the DOM tree":

``` javascript
$(document).ready(function() {
  $(document).on('click', '.boat_image', function() {
    $(this).siblings(".expedition_form_div").css("display", "inline");
  });
});
```

This works because the handler is attached once, to the document root, which stays around as Turbolinks swaps out all the content. In the way you have it, a separate click handler is attached to each link, and so you need to re-attach a new handler to links as they're loaded into the page with ajax.

Let me know if that makes sense!

PS: I also fixed your numbering in that list—markdown lists can be tricky!